+++ lib/flipper/adapters/sync/interval_synchronizer.rb
@@ -2,10 +2,10 @@ module Flipper
   module Adapters
     class Sync
       # Internal: Wraps a Synchronizer instance and only invokes it every
-      # N milliseconds.
+      # N seconds.
       class IntervalSynchronizer
-        # Private: Number of milliseconds between syncs (default: 10 seconds).
-        DEFAULT_INTERVAL_MS = 10_000
+        # Private: Number of seconds between syncs (default: 10).
+        DEFAULT_INTERVAL = 10
 
         # Private
         def self.now_ms
@@ -23,7 +23,7 @@ module Flipper
         #            the wrapped synchronizer.
         def initialize(synchronizer, interval: nil)
           @synchronizer = synchronizer
-          @interval = interval || DEFAULT_INTERVAL_MS
+          @interval = interval || DEFAULT_INTERVAL
           # TODO: add jitter to this so all processes booting at the same time
           # don't phone home at the same time.
           @last_sync_at = 0
@@ -41,7 +41,7 @@ module Flipper
         private
 
         def time_to_sync?
-          (now_ms - @last_sync_at) >= @interval
+          ((now_ms - @last_sync_at) / 1_000.0) >= @interval
         end
 
         def now_ms